% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{check_CCC_packages}
\alias{check_CCC_packages}
\title{Check CCC packages}
\usage{
check_CCC_packages(list_methods)
}
\arguments{
\item{list_methods}{List of cell-cell communication methods}
}
\value{

}
\description{
Check R and python packages of cell-cell communication methods
}
