% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_CCC_packages}
\alias{install_CCC_packages}
\title{Install CCC packages}
\usage{
install_CCC_packages(list_methods)
}
\arguments{
\item{List}{of cell-cell communication methods}
}
\description{
Install R and python packages of cell-cell communication methods
}
