% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.R
\name{run_aggregate_results}
\alias{run_aggregate_results}
\title{Aggregate CCC results}
\usage{
run_aggregate_results(name_mat, Lcell, Rcell, list_methods, path_result, top)
}
\arguments{
\item{name_mat}{Name of scRNA-seq data}

\item{Lcell}{Name of ligand(sender) cell}

\item{Rcell}{Name of receptor(receiver) cell}

\item{list_methods}{List of cell-cell communication methods}

\item{path_result}{Path of result}

\item{top}{Number of top inferred results}
}
\description{
merge the top inferred results of each method sorted by communication score or P-value
}
