% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cell_communication_methods.R
\name{run_CCC_methods}
\alias{run_CCC_methods}
\title{Run CCC methods}
\usage{
run_CCC_methods(
  name_mat,
  Lcell,
  Rcell,
  mat_ori,
  label_ori,
  list_methods,
  path_result
)
}
\arguments{
\item{name_mat}{Name of scRNA-seq data}

\item{Lcell}{Name of ligand(sender) cell}

\item{Rcell}{Name of receptor(receiver) cell}

\item{mat_ori}{Original expression matrix}

\item{label_ori}{Original cell label}

\item{list_methods}{List of cell-cell communication methods}

\item{path_result}{Path of result}
}
\description{
Run cell-cell communication methods
}
