% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cell_communication_methods.R
\name{run_CCC_methods_simulated}
\alias{run_CCC_methods_simulated}
\title{Run CCC methods in simulated data}
\usage{
run_CCC_methods_simulated(
  list_simulated_type,
  name_mat_CTP,
  Lcell,
  Rcell,
  list_methods,
  path_data,
  path_result
)
}
\arguments{
\item{list_simulated_type}{List of simulated type}

\item{name_mat_CTP}{Name of scRNA-seq data for cell type pairs}

\item{Lcell}{Name of ligand(sender) cell}

\item{Rcell}{Name of receptor(receiver) cell}

\item{list_methods}{List of cell-cell communication methods}

\item{path_data}{Path of data}

\item{path_result}{Path of result}
}
\value{

}
\description{
Load simulated data and run cell-cell communication methods in simulated data
}
