% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{run_convert_gene_symbol_mouse2human}
\alias{run_convert_gene_symbol_mouse2human}
\title{Convert gene symbol}
\usage{
run_convert_gene_symbol_mouse2human(path_data, name_mat, species)
}
\arguments{
\item{path_data}{Path of data}

\item{name_mat}{Name of scRNA-seq data}

\item{species}{Species name of scRNA-seq data}
}
\value{

}
\description{
Convert mouse gene symbol to mouse gene symbol
}
