% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_robustness.R
\name{run_evaluate_robustness}
\alias{run_evaluate_robustness}
\title{Evaluate robustness}
\usage{
run_evaluate_robustness(
  path_result,
  list_simulated_type,
  list_methods,
  name_mat,
  Lcell,
  Rcell,
  name_mat_1 = NULL,
  name_mat_2 = NULL
)
}
\arguments{
\item{path_result}{Path of result}

\item{list_simulated_type}{List of simulated type}

\item{list_methods}{List of cell-cell communication methods}

\item{name_mat}{Name of scRNA-seq data}

\item{Lcell}{Name of ligand(sender) cell}

\item{Rcell}{Name of receptor(receiver) cell}

\item{name_mat_1}{Name of the first scRNA-seq replicate data}

\item{name_mat_2}{Name of the second scRNA-seq replicate data}
}
\value{

}
\description{
Evaluate robustness of cell-cell communication by calculating the similarity
between inferred LR pairs in the original data and simulated data. The name_mat_1
and name_mat_2 are needed if data_type=='bioReplicate'
}
