% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulated_data.R
\name{run_generate_simulated_data}
\alias{run_generate_simulated_data}
\title{Generate simulated data}
\usage{
run_generate_simulated_data(
  list_simulated_type,
  path_data,
  name_mat_CTP,
  mat_ori,
  name_label_CTP,
  label_ori,
  Lcell,
  Rcell
)
}
\arguments{
\item{list_simulated_type}{List of simulated type}

\item{path_data}{Path of data}

\item{name_mat_CTP}{Name of scRNA-seq data for cell type pairs}

\item{mat_ori}{Original expression matrix}

\item{name_label_CTP}{Name of cell label for cell type pairs}

\item{label_ori}{Original cell label}

\item{Lcell}{Name of ligand(sender) cell}

\item{Rcell}{Name of receptor(receiver) cell}
}
\description{
Generate simulated data
}
