% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{run_split_mat_to_cell_type_pairs}
\alias{run_split_mat_to_cell_type_pairs}
\title{Split original data}
\usage{
run_split_mat_to_cell_type_pairs(path_data, name_mat, name_label, Lcell, Rcell)
}
\arguments{
\item{path_data}{Path of data}

\item{name_mat}{Name of scRNA-seq data}

\item{name_label}{Name of cell label}

\item{Lcell}{Name of ligand(sender) cell}

\item{Rcell}{Name of receptor(receiver) cell}
}
\description{
Split original scRNA-seq data to cell type pair scRNA-seq data
}
